Cleanup a few helpers

+++ src/main/java/net/coopfury/JukeItOut/helpers/java/TimeUnits.java
@@ -10,7 +10,7 @@ public enum TimeUnits {
         this.unitMultiplier = unitMultiplier;
     }
 
-    public long encode(int val) {
+    public long encode(long val) {
         return val * unitMultiplier;
     }
